refactor(material/menu): move trigger logic into base class
Moves most of the menu trigger logic into a base class so it can be reused for the context menu.
