+++ src/material/menu/BUILD.bazel
@@ -76,6 +76,7 @@ ng_project(
         "menu-panel.ts",
         "menu-positions.ts",
         "menu-trigger.ts",
+        "menu-trigger-base.ts",
         "module.ts",
         "public-api.ts",
     ],
